fixed misuses of NSDebugLog

+++ Source/x11/XGServerWindow.m
@@ -2303,7 +2303,7 @@ _get_next_prop_new_event(Display *display, XEvent *event, char *arg)
       return;
     }
 
-  NSDebugLLog(@"XGTrace", @"DPStermwindow: %d", win);
+  NSDebugLLog(@"XGTrace", @"DPStermwindow: %ld", (long)win);
   if (window->ic)
     {
       [inputServer ximCloseIC: window->ic];
@@ -2480,7 +2480,7 @@ NSLog(@"styleoffsets ... guessing offsets\n");
   if (!window)
     return;
 
-  NSDebugLLog(@"XGTrace", @"DPSstylewindow: %d : %d", style, win);
+  NSDebugLLog(@"XGTrace", @"DPSstylewindow: %lu : %ld", (unsigned long)style, (long)win);
   if (window->win_attrs.window_style != style
     || (window->win_attrs.flags & GSWindowStyleAttr) == 0)
     {
@@ -2527,7 +2527,7 @@ NSLog(@"styleoffsets ... guessing offsets\n");
   xf.red   = 65535 * [color redComponent];
   xf.green = 65535 * [color greenComponent];
   xf.blue  = 65535 * [color blueComponent];
-  NSDebugLLog(@"XGTrace", @"setbackgroundcolor: %@ %d", color, win);
+  NSDebugLLog(@"XGTrace", @"setbackgroundcolor: %@ %ld", color, (long)win);
   xf = [self xColorFromColor: xf forScreen: window->screen];
   window->xwn_attrs.background_pixel = xf.pixel;
   XSetWindowBackground(dpy, window->ident, window->xwn_attrs.background_pixel);
@@ -2541,7 +2541,7 @@ NSLog(@"styleoffsets ... guessing offsets\n");
   if (!window)
     return;
 
-  NSDebugLLog(@"XGTrace", @"DPSwindowbacking: %d : %d", (int)type, win);
+  NSDebugLLog(@"XGTrace", @"DPSwindowbacking: %ld : %ld", (long)type, (long)win);
 
   if ((window->gdriverProtocol & GDriverHandlesBacking))
     {
@@ -2621,7 +2621,7 @@ NSLog(@"styleoffsets ... guessing offsets\n");
   if (!window)
     return;
 
-  NSDebugLLog(@"XGTrace", @"DPSdocedited: %d : %d", edited, win);
+  NSDebugLLog(@"XGTrace", @"DPSdocedited: %ld : %ld", (long)edited, (long)win);
   window->win_attrs.flags |= GSExtraFlagsAttr;
   if (edited)
     {
@@ -2669,7 +2669,7 @@ NSLog(@"styleoffsets ... guessing offsets\n");
     {
       return;
     }
-  NSDebugLLog(@"XGTrace", @"DPSminiwindow: %d ", win);
+  NSDebugLLog(@"XGTrace", @"DPSminiwindow: %ld ", (long)win);
   /*
    * If we haven't already done so - set the icon window hint for this
    * window so that the GNUstep miniwindow is displayed (if supported).
@@ -2727,7 +2727,7 @@ NSLog(@"styleoffsets ... guessing offsets\n");
   gswindow_device_t *window;
   CGFloat t, b, l, r;
 
-  NSDebugLLog(@"XGTrace", @"DPSwindowdevice: %d ", win);
+  NSDebugLLog(@"XGTrace", @"DPSwindowdevice: %ld ", (long)win);
   window = WINDOW_WITH_TAG(win);
   if (!window)
     {
@@ -2931,7 +2931,7 @@ static BOOL didCreatePixmaps;
 	}
     }
 
-  NSDebugLLog(@"XGTrace", @"DPSorderwindow: %d : %d : %d",op,otherWin,winNum);
+  NSDebugLLog(@"XGTrace", @"DPSorderwindow: %ld : %ld : %ld", (long)op, (long)otherWin, (long)winNum);
   level = window->win_attrs.window_level;
   if (otherWin > 0)
     {
@@ -3182,8 +3182,7 @@ static BOOL didCreatePixmaps;
       return;
     }
 
-  NSDebugLLog(@"XGTrace", @"DPSplacewindow: %@ : %d", NSStringFromRect(rect), 
-	      win);
+  NSDebugLLog(@"XGTrace", @"DPSplacewindow: %@ : %ld", NSStringFromRect(rect), (long)win);
   nswin  = GSWindowWithNumber(win);
   frame = [nswin frame];
   if (NSEqualRects(rect, frame) == YES)
@@ -3285,7 +3284,7 @@ static BOOL didCreatePixmaps;
   if (!window)
     return NSZeroRect;
 
-  NSDebugLLog(@"XGTrace", @"DPScurrentwindowbounds: %d", win);
+  NSDebugLLog(@"XGTrace", @"DPScurrentwindowbounds: %ld", (long)win);
 
   // get the current xframe of the window
   XGetGeometry(dpy, window->ident, &window->root,
@@ -3306,7 +3305,7 @@ static BOOL didCreatePixmaps;
   if (!window)
     return;
 
-  NSDebugLLog(@"XGTrace", @"DPSsetwindowlevel: %d : %d", level, win);
+  NSDebugLLog(@"XGTrace", @"DPSsetwindowlevel: %ld : %ld", (long)level, (long)win);
   if ((int)(window->win_attrs.window_level) != level
     || (window->win_attrs.flags & GSWindowLevelAttr) == 0)
     {
@@ -3667,8 +3666,8 @@ static BOOL didCreatePixmaps;
       return;
     }
 
-  NSDebugLLog(@"XGFlush", @"DPSflushwindowrect: %@ : %d", 
-	      NSStringFromRect(rect), win);
+  NSDebugLLog(@"XGFlush", @"DPSflushwindowrect: %@ : %ld",
+	      NSStringFromRect(rect), (long)win);
   if (window->type == NSBackingStoreNonretained)
     {
       XFlush(dpy);
@@ -3788,11 +3787,11 @@ static BOOL didCreatePixmaps;
 		     GrabModeAsync, GrabModeAsync, None, None, [self lastTime]);
 
   if (ret != GrabSuccess)
-    NSDebugLLog(@"XGTrace", @"Failed to grab pointer %d\n", win);
+    NSDebugLLog(@"XGTrace", @"Failed to grab pointer %ld\n", (long)win);
   else
     {
       grab_window = window;
-      NSDebugLLog(@"XGTrace", @"Grabbed pointer %d\n", win);
+      NSDebugLLog(@"XGTrace", @"Grabbed pointer %ld\n", (long)win);
     }
   return (ret == GrabSuccess) ? YES : NO;
 }
@@ -3823,11 +3822,11 @@ static BOOL didCreatePixmaps;
 
   if (win == 0 || window == 0)
     {
-      NSDebugLLog(@"Focus", @"Setting focus to unknown win %d", win);
+      NSDebugLLog(@"Focus", @"Setting focus to unknown win %ld", (long)win);
       return;
     }
 
-  NSDebugLLog(@"XGTrace", @"DPSsetinputfocus: %d", win);
+  NSDebugLLog(@"XGTrace", @"DPSsetinputfocus: %ld", (long)win);
   /*
    * If we have an outstanding request to set focus to this window,
    * we don't want to do it again.
@@ -3869,7 +3868,7 @@ static BOOL didCreatePixmaps;
   if (!handlesWindowDecorations)
     return;
 
-  NSDebugLLog(@"XGTrace", @"DPSsetinputstate: %d : %d", st, win);
+  NSDebugLLog(@"XGTrace", @"DPSsetinputstate: %ld : %ld", (long)st, (long)win);
   if ((generic.wm & XGWM_WINDOWMAKER) != 0)
     {
       gswindow_device_t *window = WINDOW_WITH_TAG(win);
@@ -3930,11 +3929,11 @@ static BOOL didCreatePixmaps;
 
   if (win == 0 || window == 0)
     {
-      NSDebugLLog(@"XGTrace", @"Setting alpha to unknown win %d", win);
+      NSDebugLLog(@"XGTrace", @"Setting alpha to unknown win %ld", (long)win);
       return;
     }
 
-  NSDebugLLog(@"XGTrace", @"setalpha: %d", win);
+  NSDebugLLog(@"XGTrace", @"setalpha: %ld", (long)win);
   
   /* Initialize the atom if needed */
   if (opacity_atom == None)
@@ -3974,7 +3973,7 @@ static BOOL didCreatePixmaps;
 
   if (win == 0 || window == 0)
     {
-      NSDebugLLog(@"XGTrace", @"Setting alpha to unknown win %d", win);
+      NSDebugLLog(@"XGTrace", @"Setting alpha to unknown win %ld", (long)win);
       return alpha;
     }
